added site, project, and site_project models, controllers, specs

+++ db/migrate/20161229013302_create_hubs.rb
@@ -2,7 +2,7 @@ class CreateHubs < ActiveRecord::Migration[5.0]
   def change
     create_table :hubs, id: :uuid do |t|
       t.string :name
-      t.float :percent_fee
+      t.decimal :percent_fee
 
       t.timestamps
     end
